Refactors inspectors within components Refactors the InspectorButton into the InspectorPanel Adds tooltips to the buttons
