+++ v3/src/components/case-table/case-table-inspector.tsx
@@ -1,47 +1,33 @@
-import React, { ReactNode } from "react"
-import { Button } from "@chakra-ui/react"
-import { InspectorPanel } from "../inspector-panel"
+import React from "react"
+import { InspectorButton, InspectorPanel } from "../inspector-panel"
 import HideShowIcon from "../../assets/icons/icon-hideShow.svg"
 import InformationIcon from "../../assets/icons/icon-info.svg"
 import ScaleDataIcon from "../../assets/icons/icon-scaleData.svg"
 import TrashIcon from "../../assets/icons/icon-trash.svg"
 import ValuesIcon from "../../assets/icons/icon-values.svg"
-import MoreOptionsIcon from "../../assets/icons/arrow-moreIconOptions.svg"
 
 interface IProps {
   show: boolean
 }
 
-const ToolIconMap: Record<string, ReactNode> = {
-  "information": <InformationIcon />,
-  "resize": <ScaleDataIcon />,
-  "trash": <TrashIcon />,
-  "hide_show": <HideShowIcon />,
-  "values": <ValuesIcon />
-}
-
 export const CaseTableInspector = ({ show }: IProps) => {
   return (show
     ? <InspectorPanel>
-        <InspectorButton type={"information"} />
-        <InspectorButton type={"resize"} />
-        <InspectorButton type={"trash"} />
-        <InspectorButton type={"hide_show"} />
-        <InspectorButton type={"values"} />
+        <InspectorButton icon={<InformationIcon />} type={"datasetInfo"} />
+        <InspectorButton icon={<ScaleDataIcon />} type={"resize"} />
+        <InspectorButton icon={<TrashIcon />} type={"delete"} />
+        <InspectorButton icon={<HideShowIcon />} type={"hideShow"} />
+        <InspectorButton icon={<ValuesIcon />} type={"attributes"} />
       </InspectorPanel>
     : null
   )
 }
 
-interface IInspectorButtonProps {
-  type: string
-}
+// const [showInfoModal, setShowInfoModal] = useState(false)
 
-const InspectorButton = ({type}:IInspectorButtonProps) => {
-  return (
-    <Button className="inspector-tool-button">
-      {ToolIconMap[type]}
-      {type !== "resize" && <MoreOptionsIcon className="more-options-icon"/>}
-    </Button>
-  )
-}
+// const handleToolClick = (tool: string) => {
+//   switch (tool) {
+//     case "datasetInfo":
+//       setShowInfoModal(true)
+//   }
+// }
