+++ v3/src/components/graph/components/graph-component.tsx
@@ -11,7 +11,6 @@ import {DataConfigurationModel} from "../models/data-configuration-model"
 import {GraphLayout, GraphLayoutContext} from "../models/graph-layout"
 import {GraphModel} from "../models/graph-model"
 import {Graph} from "./graph"
-import { GraphInspector } from './graph-inspector'
 
 const defaultGraphModel = GraphModel.create({
   axes: {
@@ -33,7 +32,7 @@ export const GraphComponent = observer(({broker}: IProps) => {
   const enableAnimation = useRef(true)
   const dataset = broker?.selectedDataSet || broker?.last
   const dotsRef = useRef<SVGSVGElement>(null)
-  const [showInspector, setShowInspector] = useState(true)
+  const [showInspector, setShowInspector] = useState(false)
 
   useEffect(() => {
     (width != null) && (height != null) && layout.setGraphExtent(width, height)
@@ -55,7 +54,6 @@ export const GraphComponent = observer(({broker}: IProps) => {
               showInspector={showInspector}
               setShowInspector={setShowInspector}
             />
-            <GraphInspector show={showInspector} />
           </GraphLayoutContext.Provider>
         </InstanceIdContext.Provider>
       </DataSetContext.Provider>
