+++ v3/src/components/inspector-panel.scss
@@ -1,4 +1,4 @@
-@import "../vars.scss";
+@import "./vars.scss";
 
 .inspector-panel {
   position: absolute;
