--- v3/src/components/inspector-panel/inspector-panel.tsx
@@ -1,79 +1,0 @@
-import { Box, Button } from "@chakra-ui/react"
-import React, { useState } from "react"
-import InformationIcon from "../../assets/icons/icon-info.svg"
-import LayersIcon from "../../assets/icons/icon-layers.svg"
-import ScaleDataIcon from "../../assets/icons/icon-scaleData.svg"
-import BarChartIcon from "../../assets/icons/icon-segmented-bar-chart.svg"
-import StylesIcon from "../../assets/icons/icon-styles.svg"
-import TrashIcon from "../../assets/icons/icon-trash.svg"
-import ValuesIcon from "../../assets/icons/icon-values.svg"
-import HideShowIcon from "../../assets/icons/icon-hideShow.svg"
-import CameraIcon from "../../assets/icons/icon-camera.svg"
-import MoreOptionsIcon from "../../assets/icons/arrow-moreIconOptions.svg"
-import t from "../../utilities/translation/translate"
-
-import "./inspector-panel.scss"
-import { DatasetInfoModal } from "./dataset-info-modal"
-
-interface IProps {
-  component: string
-}
-
-const InspectorIconComponent: Record<string, any> = {
-  "datasetInfo": <InformationIcon />,
-  "layers": <LayersIcon />,
-  "resize": <ScaleDataIcon />,
-  "bar_chart": <BarChartIcon />,
-  "styles": <StylesIcon />,
-  "trash": <TrashIcon />,
-  "values": <ValuesIcon />,
-  "hide_show": <HideShowIcon />,
-  "snapshot": <CameraIcon />
-}
-
-export const InspectorTool = (tool: string) => {
-  const IconComponent = InspectorIconComponent[tool]
-  return IconComponent
-}
-
-export const InspectorPanel = ({component}: IProps) => {
-  const tableInpectors = ["datasetInfo", "resize", "trash", "hide_show", "values"]
-  const graphInspectors = ["resize", "hide_show", "values", "bar_chart", "styles", "snapshot"]
-  const mapInspectors = ["resize", "hide_show", "values", "layers", "snapshot"]
-  const sliderInspectors = ["values"]
-  const inspectorMap: Record<string, any> = {
-    "table": tableInpectors,
-    "graph": graphInspectors,
-    "map": mapInspectors,
-    "slider": sliderInspectors
-  }
-  const [showInfoModal, setShowInfoModal] = useState(false)
-
-  const handleToolClick = (tool: string) => {
-    switch (tool) {
-      case "datasetInfo":
-        setShowInfoModal(true)
-    }
-  }
-
-  return (
-    <>
-      <Box className={`inspector-panel ${component}`} bg="tealDark" data-testid={`inspector-panel-${component}`}>
-        {inspectorMap[component].map((iType: string) => {
-          return (
-            <Button key={iType} className="inspector-tool-button" bg="tealDark"
-              title={t(`DG.Inspector.${iType}.toolTip`)} onClick={()=> handleToolClick(iType)}>
-              {InspectorTool(iType)}
-              {!(iType === "resize") &&
-                <MoreOptionsIcon className="more-options-icon"/>
-              }
-            </Button>
-          )
-        })}
-      </Box>
-    `{showInfoModal &&
-      <DatasetInfoModal showInfoModal={showInfoModal} setShowInfoModal={setShowInfoModal} />}
-    </>
-
-  )
-}
